+++ src/Nodes/Led/Led.hpp
@@ -14,7 +14,7 @@ class LedObject {
     friend SmartIotLed;
     friend SmartIotInternals::LedPattern;
     public:
-        LedObject(const uint8_t firstPos,const uint8_t nbLed,const char* name);
+        LedObject(const uint16_t firstPos,const uint16_t nbLed,const char* name);
         void display();
         void dimAll(byte value);
         void setColor(CRGB color);
@@ -40,12 +40,12 @@ class LedObject {
         uint16_t getAudio() const;
 
     protected:
-        uint8_t _nbLed;
+        uint16_t _nbLed;
         CRGB* _leds;
 
     private:
         const char* _name;
-        uint8_t _firstPos;
+        uint16_t _firstPos;
         uint8_t _speed;
         bool _state;
         bool _autoplay;
